create ProcessPokemonMap
ProcessPokemonMap creates a map with string keys and list of string
values where the keys are pokemon regions and the list of strings
contain the strings that have the pokemon IDs and names. These will need
to be split into their own lists, creating a map with regions as keys
and a list of lists of strings containing the ID and name. The function
will be changed to that next.