update ProcessPokemonMap
Added splitStrInList helper function (and popFirst forgot to mention
before) to split the string containing the ID and name in a list that is
added to a matrix and returned. ProcessPokemonMap uses helper function
to change the list of pokemon strings into a matrix then save that as a
value where the key is the region of all pokemon in the matrix. This
will be used to setup Pokemon structs later once type is extracted and
added to the map.