parse for pokemon type page urls
This required getting the parent nodes of the <img> tags with class
attributes equal to "typeimg". The parent nodes are <a> tags with href
attributes to each type page. Created a function that gets the parent
nodes of a specific element (GetDOMParentNode). Then parse and create
urls for each type page and return a list of the urls
(ParseTypePageUrls).